♻️ Extract abstract class for ObjectFactory
In constructing object inheritance, a more robust strategy is to create
an abstract class and then have classes directly extend that abstract
class.

This helps define and narrow an interface.
